i18n + Petit rangement + ajout .env
