+++ .env.example
@@ -1,0 +1,5 @@
+DEBUG=False
+LIST_SERVERS=False
+TOKEN="TOKEN"
+PREFIX="$"
+LANGUAGE="EN"
