+++ .gitignore
@@ -1,3 +1,3 @@
 PrivateBot.png
 __pycache__
-config.json
+.env
