--- cogs/FR/kick.py
@@ -1,25 +1,0 @@
-import discord
-from discord.ext import commands
-class Kick(commands.Cog):
-    def __init__(self, bot):
-        super().__init__()
-        self.bot = bot
-    @commands.command(name="kick",description="Permet de kicker un membre")
-    async def kick(self, ctx, member: discord.Member=None, *, reason="Aucune raison fournie"):
-        prefix = self.bot.command_prefix
-        if not ctx.author.guild_permissions.kick_members:
-            return await ctx.send("Vous n'avez pas la permission d'expulser un membre")
-        if not member:
-            return await ctx.reply(f"Synthaxe: `{prefix}kick <membre> [<raison>]`")
-        try:
-            await member.kick(reason=reason)
-        except:
-            return await ctx.reply("Je n'ai pas la permission d'expulser ce membre")
-        await ctx.send(f"{member} a bien été expulsé")
-        try:
-            await member.send(f"Tu as été expulsé de {ctx.guild.name} pour {reason} par {ctx.author}")
-        except:
-            pass
-
-async def setup(bot):
-    await bot.add_cog(Kick(bot))
