--- cogs/FR/on_ready.py
@@ -1,39 +1,0 @@
-import discord
-from discord.ext import commands
-import json
-class OnReady(commands.Cog):
-    def __init__(self,bot):
-        super().__init__()
-        self.bot = bot
-        
-    @commands.Cog.listener()
-    async def on_ready(self):
-        print(f"Connectés en tant que: {self.bot.user}({self.bot.user.id})")
-        print("Je suis sur les serveurs suivants:")
-        for guild in self.bot.guilds:
-            print(f"{guild.name} ({guild.id})")
-        with open("config.json","r",encoding="utf-8") as f:
-            config = json.load(f)
-        if config['statut']['playing'] == "y":
-            if config['statut']['text']:
-                await self.bot.change_presence(activity=discord.Game(name=config['statut']['text']))
-                return
-        if config['statut']['watching'] == "y":
-            if config['statut']['text']:
-                await self.bot.change_presence(activity=discord.Activity(type=discord.ActivityType.watching, name=config['statut']['text']))
-                return
-        if config['statut']['listening'] == "y":
-            if config['statut']['text']:
-                await self.bot.change_presence(activity=discord.Activity(type=discord.ActivityType.listening, name=config['statut']['text']))
-
-                return
-        if config['statut']['streaming'] == "y":
-            if config['statut']['url']:
-                if config['statut']['text']:
-                    await self.bot.change_presence(activity=discord.Streaming(name=config['statut']['text'], url=config['statut']['url']))
-                    return
-        print("Prêt à aider !")
-
-
-async def setup(bot):
-    await bot.add_cog(OnReady(bot))
