+++ config.json
@@ -1,8 +1,4 @@
 {
-    "debug": false,
-    "token": "TOKEN",
-    "prefix": "$",
-    "lang": "fr",
     "statut": {
         "playing": "n",
         "watching": "y",
@@ -15,7 +11,6 @@
     "owner": [
         
     ],
-    
     "logs": {
         "activate": "n",
         "channel": "channel ID"
