+++ locales/EN/Events.json
@@ -1,0 +1,3 @@
+{
+    "CONNECTED_AS": "Connected As: %{ME}"
+}
