+++ locales/FR/Events.json
@@ -1,0 +1,3 @@
+{
+    "CONNECTED_AS": "Connecté en tant que: %{ME}"
+}
