+++ main.py
@@ -1,44 +1,45 @@
 import json
 import os
-
-import traceback
 import sys
+import traceback
+
+from dotenv import load_dotenv
 
 import discord
 from discord.ext import commands
+import i18n
 
-with open("config.json") as f:
-    config = json.load(f)
+load_dotenv()
 
-lang = config['lang'].upper()
-list_lang = []
+i18n.load_path.append("locales")
+i18n.set("file_format", "json")
+i18n.set("filename_format", "{namespace}.{format}")
+i18n.set("namespace_delimiter", ":")
+i18n.set("skip_locale_root_data", True)
+i18n.set("use_locale_dirs", True)
 
-for folder in os.listdir("./cogs/"):
-    folder = folder.upper()
-    list_lang.append(folder)
+DEBUG = os.environ["DEBUG"] == "True"
 
 class PrivateBot(commands.Bot):
     def __init__(self) -> None:
-        super().__init__(command_prefix=commands.when_mentioned_or(config['prefix']),intents=discord.Intents.all())
+        super().__init__(command_prefix=commands.when_mentioned_or(os.environ['PREFIX']),intents=discord.Intents.all())
         self.remove_command("help")
 
     async def LoadCogs(self):
-        if lang in list_lang:
-                for file in os.listdir(f"./cogs/{lang}"):
-                    if file.endswith(".py"):
-                        cog = f"cogs.{lang}.{file[:-3]}"
-                        try:
-                            await bot.load_extension(cog)
-                            if(config["debug"]):
-                                print(f'Loaded {cog} !', file=sys.stdout)
-                        except Exception as e:
-                            print(f'Failed to load {cog}.', file=sys.stderr)
-                            if(config["debug"]):
-                                traceback.print_exc()
+        for file in os.listdir(f"./cogs"):
+            if file.endswith(".py"):
+                cog = f"cogs.{file[:-3]}"
+                try:
+                    await bot.load_extension(cog)
+                    if(DEBUG):
+                        print(f'Loaded {cog} !', file=sys.stdout)
+                except Exception as e:
+                    print(f'Failed to load {cog}.', file=sys.stderr)
+                    if(DEBUG):
+                        traceback.print_exc()
 
     async def setup_hook(self) -> None:
         await self.LoadCogs()
 
 bot = PrivateBot()
-
-bot.run(config['token'])
+bot.run(os.environ['TOKEN'])
