+++ requirements.txt
@@ -1,1 +1,3 @@
-discord.py==1.7.3
+discord.py==2.3.0
+i18nice
+python-dotenv
